refactor(pug-parser): move codes to types.ts
